Backend/api/requirements.txt
numpy
pandas
uvicorn
fastapi
passlib
python-jose
python-dotenv
snowflake-connector-python
bcrypt
python-multipart
jose